+++ lualib-src/lua-sharedata.c
@@ -1,25 +1,26 @@
-// build: gcc -O2 -Wall --shared -o conf.so luaconf.c 
-
 #include <lua.h>
 #include <lauxlib.h>
 #include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 #include <assert.h>
+#include "atomic.h"
 
 #define KEYTYPE_INTEGER 0
 #define KEYTYPE_STRING 1
 
 #define VALUETYPE_NIL 0
-#define VALUETYPE_NUMBER 1
+#define VALUETYPE_REAL 1
 #define VALUETYPE_STRING 2
 #define VALUETYPE_BOOLEAN 3
 #define VALUETYPE_TABLE 4
+#define VALUETYPE_INTEGER 5
 
 struct table;
 
 union value {
 	lua_Number n;
+	lua_Integer d;
 	struct table * tbl;
 	int string;
 	int boolean;
@@ -71,11 +72,10 @@ countsize(lua_State *L, int sizearray) {
 		int type = lua_type(L, -2);
 		++n;
 		if (type == LUA_TNUMBER) {
-			lua_Number key = lua_tonumber(L, -2);
-			int nkey = (int)key;
-			if ((lua_Number)nkey != key) {
-				luaL_error(L, "Invalid key %f", key);
+			if (!lua_isinteger(L, -2)) {
+				luaL_error(L, "Invalid key %f", lua_tonumber(L, -2));
 			}
+			lua_Integer nkey = lua_tointeger(L, -2);
 			if (nkey > 0 && nkey <= sizearray) {
 				--n;
 			}
@@ -129,8 +129,13 @@ setvalue(struct context * ctx, lua_State *L, int index, struct node *n) {
 		n->valuetype = VALUETYPE_NIL;
 		break;
 	case LUA_TNUMBER:
-		n->v.n = lua_tonumber(L, index);
-		n->valuetype = VALUETYPE_NUMBER;
+		if (lua_isinteger(L, index)) {
+			n->v.d = lua_tointeger(L, index);
+			n->valuetype = VALUETYPE_INTEGER;
+		} else {
+			n->v.n = lua_tonumber(L, index);
+			n->valuetype = VALUETYPE_REAL;
+		}
 		break;
 	case LUA_TSTRING: {
 		size_t sz = 0;
@@ -372,6 +377,7 @@ pconv(lua_State *L) {
 static void
 convert_stringmap(struct context *ctx, struct table *tbl) {
 	lua_State *L = ctx->L;
+	lua_checkstack(L, ctx->string_index + LUA_MINSTACK);
 	lua_settop(L, ctx->string_index + 1);
 	lua_pushvalue(L, 1);
 	struct state * s = lua_newuserdata(L, sizeof(*s));
@@ -468,8 +474,11 @@ ldeleteconf(lua_State *L) {
 static void
 pushvalue(lua_State *L, lua_State *sL, uint8_t vt, union value *v) {
 	switch(vt) {
-	case VALUETYPE_NUMBER:
+	case VALUETYPE_REAL:
 		lua_pushnumber(L, v->n);
+		break;
+	case VALUETYPE_INTEGER:
+		lua_pushinteger(L, v->d);
 		break;
 	case VALUETYPE_STRING: {
 		size_t sz = 0;
@@ -530,16 +539,15 @@ lindexconf(lua_State *L) {
 	size_t sz = 0;
 	const char * str = NULL;
 	if (kt == LUA_TNUMBER) {
-		lua_Number k = lua_tonumber(L, 2);
-		key = (int)k;
-		if ((lua_Number)key != k) {
-			return luaL_error(L, "Invalid key %f", k);
+		if (!lua_isinteger(L, 2)) {
+			return luaL_error(L, "Invalid key %f", lua_tonumber(L, 2));
 		}
+		key = (int)lua_tointeger(L, 2);
 		if (key > 0 && key <= tbl->sizearray) {
 			--key;
 			pushvalue(L, tbl->L, tbl->arraytype[key], &tbl->array[key]);
 			return 1;
-		} 
+		}
 		keytype = KEYTYPE_INTEGER;
 		keyhash = (uint32_t)key;
 	} else {
@@ -601,13 +609,12 @@ lnextkey(lua_State *L) {
 	const char *str = NULL;
 	int sizearray = tbl->sizearray;
 	if (kt == LUA_TNUMBER) {
-		lua_Number k = lua_tonumber(L,2);
-		key = (int)k;
-		if ((lua_Number)key != k) {
+		if (!lua_isinteger(L, 2)) {
 			return 0;
 		}
+		key = (int)lua_tointeger(L, 2);
 		if (key > 0 && key <= sizearray) {
-			int i;
+			lua_Integer i;
 			for (i=key;i<sizearray;i++) {
 				if (tbl->arraytype[i] != VALUETYPE_NIL) {
 					lua_pushinteger(L, i+1);
@@ -657,7 +664,7 @@ releaseobj(lua_State *L) {
 	struct ctrl *c = lua_touserdata(L, 1);
 	struct table *tbl = c->root;
 	struct state *s = lua_touserdata(tbl->L, 1);
-	__sync_fetch_and_sub(&s->ref, 1);
+	ATOM_DEC(&s->ref);
 	c->root = NULL;
 	c->update = NULL;
 
@@ -668,7 +675,7 @@ static int
 lboxconf(lua_State *L) {
 	struct table * tbl = get_table(L,1);	
 	struct state * s = lua_touserdata(tbl->L, 1);
-	__sync_fetch_and_add(&s->ref, 1);
+	ATOM_INC(&s->ref);
 
 	struct ctrl * c = lua_newuserdata(L, sizeof(*c));
 	c->root = tbl;
@@ -713,7 +720,7 @@ static int
 lincref(lua_State *L) {
 	struct table *tbl = get_table(L,1);
 	struct state * s = lua_touserdata(tbl->L, 1);
-	int ref = __sync_add_and_fetch(&s->ref, 1);
+	int ref = ATOM_INC(&s->ref);
 	lua_pushinteger(L , ref);
 
 	return 1;
@@ -723,7 +730,7 @@ static int
 ldecref(lua_State *L) {
 	struct table *tbl = get_table(L,1);
 	struct state * s = lua_touserdata(tbl->L, 1);
-	int ref = __sync_sub_and_fetch(&s->ref, 1);
+	int ref = ATOM_DEC(&s->ref);
 	lua_pushinteger(L , ref);
 
 	return 1;