+++ service/cdummy.lua
@@ -1,8 +1,10 @@
 local skynet = require "skynet"
+require "skynet.manager"	-- import skynet.launch, ...
 
 local globalname = {}
 local queryname = {}
 local harbor = {}
+local harbor_service
 
 skynet.register_protocol {
 	name = "harbor",